requirements.txt
amqp==5.3.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pyTelegramBotAPI==4.32.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
schedule==1.2.2
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
